Conflicts:
	qemu/Makefile.target
	qemu/configure
	qemu/cpu-all.h
	qemu/cpu-exec.c
	qemu/def-helper.h
	qemu/default-configs/i386-softmmu.mak
	qemu/dyngen-exec.h
	qemu/exec-all.h
	qemu/exec.c
	qemu/hw/hw.h
	qemu/hw/pc.c
	qemu/hw/pci.h
	qemu/hw/vga.c
	qemu/monitor.c
	qemu/qemu-options.hx
	qemu/rules.mak
	qemu/softmmu_header.h
	qemu/softmmu_template.h
	qemu/target-i386/cpu.h
	qemu/target-i386/helper.c
	qemu/target-i386/helper.h
	qemu/target-i386/machine.c
	qemu/target-i386/op_helper.c
	qemu/target-i386/translate.c
	qemu/tcg/i386/tcg-target.c
	qemu/tcg/tcg.h
	qemu/vl.c

+++ qemu/cache-utils.h
@@ -2,6 +2,9 @@
 #define QEMU_CACHE_UTILS_H
 
 #if defined(_ARCH_PPC)
+
+#include <stdint.h> /* uintptr_t */
+
 struct qemu_cache_conf {
     unsigned long dcache_bsize;
     unsigned long icache_bsize;
@@ -12,7 +15,7 @@ extern struct qemu_cache_conf qemu_cache_conf;
 void qemu_cache_utils_init(char **envp);
 
 /* mildly adjusted code from tcg-dyngen.c */
-static inline void flush_icache_range(unsigned long start, unsigned long stop)
+static inline void flush_icache_range(uintptr_t start, uintptr_t stop)
 {
     unsigned long p, start1, stop1;
     unsigned long dsize = qemu_cache_conf.dcache_bsize;
